improving game components
